правки по кодревью
