+++ src/components/Layout/Layout.js
@@ -1,5 +1,7 @@
-import { Outlet, useMatch } from 'react-router-dom'
-import st from './Layout.module.scss'
+import { Outlet, useMatch } from 'react-router-dom';
+import React, { useEffect } from "react";
+import st from './Layout.module.scss';
+import { useLocation } from 'react-router-dom';
 
 import Header from '../../components/Header';
 import Footer from '../../components/Footer';
@@ -7,13 +9,29 @@ import Container from '../Container';
 
 const Layout = () => {
     const match = useMatch({ path: '/' });
-    
+    const { pathname, hash } = useLocation();
+
+    useEffect(() => {
+        if (hash) {
+            setTimeout(() => {
+                let titleScrollTo = document.getElementById(hash);
+
+                titleScrollTo && titleScrollTo.scrollIntoView({
+                    block: 'center',
+                    behavior: 'smooth'
+                });
+            }, 2000)
+        } else {
+            window.scrollTo(0, 0);
+        }
+    }, [pathname, hash]);
+
     return (
         <div className={st.root}>
             <Header />
             <main className={st.main}>
                 {
-                    match !== null ? <Outlet /> :
+                    match ? <Outlet /> :
                         <Container>
                             {/* Outlet = {children} */}
                             <Outlet />
