+++ src/pages/Biography/Biography.js
@@ -7,7 +7,6 @@ import Heading from '../../components/Heading';
 import Text from '../../components/Text';
 import Button from '../../components/Button';
 import { BIO } from '../../consts/BIO';
-import { ReactComponent as LinkIcon } from './assets/link_icon.svg';
 
 import st from './Biography.module.scss'
 
@@ -25,15 +24,6 @@ const Biography = () => {
         navigate(-1); //go to prev page
     }
 
-    // setTimeout(() => {
-    //     let titleScrollTo = document.getElementById(window.location.hash);
-
-    //     titleScrollTo && titleScrollTo.scrollIntoView({
-    //         block: 'center',
-    //         behavior: 'smooth'
-    //     });
-    // }, 3000)
-
     return (
         <section className={cn(st.root)}>
             <Container>
@@ -46,28 +36,19 @@ const Biography = () => {
                 {
                     character.map((item, i) => {
                         let node;
-                        let lvl;
-                        let anchorLink;
-
+                        
                         switch (item.type) {
                             case 'h1':
-                                lvl = item.type.replace(/[^0-9.]/g, '');
-                                node =
-                                    <Heading lvl={parseInt(lvl, 10)}>
-                                        {item.text}
-                                    </Heading>
-                                break;
-
                             case 'h2':
-                                lvl = item.type.replace(/[^0-9.]/g, '');
-                                anchorLink = item.text.replace(/\s/g, '_');
+                                let lvl = item.type.replace(/[^0-9.]/g, '');
+                                let anchorLink = item.text.replace(/\s/g, '_');
 
                                 node =
-                                    <Heading lvl={parseInt(lvl, 10)}>
+                                    <Heading 
+                                    lvl={parseInt(lvl, 10)}
+                                    anchorLink={`#${anchorLink}`}
+                                    >
                                         {item.text}
-                                        <a href={`#${anchorLink}`} id={`#${anchorLink}`}>
-                                            <LinkIcon />
-                                        </a>
                                     </Heading>
                                 break;
